+++ riscv-rt/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [Unreleased] - RT Ibex
+
+### Changed
+
+- Align stack pointer to 4-bytes on RVE
 
 ## [v0.14.0] - 2025-02-18